Initial restructure of application logic in orchest-ctl
