+++ orchest/orchest-ctl/app/errors.py
@@ -1,0 +1,7 @@
+class Error(Exception):
+    pass
+
+
+class ENVVariableNotFound(Error):
+    """ENV variable not found."""
+    pass
