+++ orchest/orchest-ctl/app/main.py
@@ -1,505 +1,34 @@
 import logging
-import os
 import sys
 
-import docker
-from docker.types import Mount
-
-from connections import docker_client
-
-
-VALID_COMMANDS = {
-    "start": "Starts the Orchest application",
-    "help": "Shows this help menu",
-    "stop": "Stops the Orchest application",
-    "status": "Checks the current status of the Orchest application",
-    "update": "Update Orchest to the latest version by pulling latest container images"
-}
-
-DOCKER_NETWORK = 'orchest'
-
-if "HOST_USER_DIR" in os.environ:
-    HOST_USER_DIR = os.environ.get("HOST_USER_DIR")
-else:
-    raise("Need to set HOST_USER_DIR!")
-
-if "HOST_CONFIG_DIR" in os.environ:
-    HOST_CONFIG_DIR = os.environ.get("HOST_CONFIG_DIR")
-else:
-    raise("Need to set HOST_CONFIG_DIR!")
-
-DURABLE_QUEUES_DIR = ".orchest/rabbitmq-mnesia"
-
-# Set to `True` if you want to pull images from Dockerhub
-# instead of using local equivalents
-
-CONTAINER_MAPPING = {
-    "orchestsoftware/orchest-api:latest": {
-        "environment": {},
-        "command": None,
-        "name": "orchest-api",
-        "mounts": [
-            {
-                # Needed for persistent db.
-                "source": HOST_USER_DIR,
-                "target": "/userdir"
-            },
-            {
-                "source": "/var/run/docker.sock",
-                "target": "/var/run/docker.sock"
-            },
-        ],
-
-    },
-    "orchestsoftware/orchest-webserver:latest": {
-        "name": "orchest-webserver",
-        "environment": {
-            "HOST_USER_DIR": HOST_USER_DIR,
-        },
-        "mounts": [
-            {
-                "source": "/var/run/docker.sock",
-                "target": "/var/run/docker.sock"
-            },
-            {
-                "source": HOST_USER_DIR,
-                "target": "/userdir"
-            },
-            {
-                "source": HOST_CONFIG_DIR,
-                "target": "/config"
-            }
-        ],
-    },
-    "orchestsoftware/celery-worker:latest": {
-        "name": "celery-worker",
-        "mounts": [
-            {
-                "source": "/var/run/docker.sock",
-                "target": "/var/run/docker.sock"
-            },
-            {
-                # Mount is needed for copying the snapshot dir to
-                # pipeline run dirs for experiments.
-                "source": HOST_USER_DIR,
-                "target": "/userdir"
-            },
-        ]
-    },
-    "rabbitmq:3": {
-        "name": "rabbitmq-server",
-        "hostname": "rabbitmq-hostname",
-        "mounts": [
-            {
-                # Persisting RabbitMQ Queues.
-                "source": os.path.join(HOST_USER_DIR, DURABLE_QUEUES_DIR),
-                "target": "/var/lib/rabbitmq/mnesia",
-            }
-        ],
-    },
-    "orchestsoftware/auth-server:latest": {
-        "name": "auth-server",
-        "environment": {},
-        "mounts": [
-            {
-                "source": HOST_CONFIG_DIR,
-                "target": "/config"
-            },
-            {
-                "source": HOST_USER_DIR,
-                "target": "/userdir"
-            },
-        ],
-    },
-    "orchestsoftware/nginx-proxy:latest": {
-        "name": "nginx-proxy",
-        "ports": {
-            "80/tcp": 8000,
-            "443/tcp": 443,
-        }
-    },
-}
-
-# TODO: shutting down can be done easier by just shutting down all the
-#       containers inside the "orchest" docker network.
-# images in CONTAINER_MAPPING need to be run in the application on start
-# additional images below are necessary for dynamically creating containers
-# in Orchest for the pipeline steps / Jupyter server / Enterprise Gateway
-
-IMAGES = list(CONTAINER_MAPPING.keys())
-IMAGES += [
-    "elyra/enterprise-gateway:2.2.0rc2",
-    "orchestsoftware/jupyter-server:latest",
-    "orchestsoftware/custom-base-kernel-py:latest",
-    "orchestsoftware/custom-base-kernel-r:latest",
-    "orchestsoftware/memory-server:latest",
-    "orchestsoftware/orchest-ctl:latest",
-]
-
-
-def clean_containers():
-    running_containers = docker_client.containers.list(all=True)
-
-    for container in running_containers:
-        has_tags = len(container.image.tags) > 0
-        is_orchest_image = container.image.tags[0] in IMAGES
-        has_exited = container.status == "exited"
-        if has_tags and is_orchest_image and has_exited:
-            logging.info("Removing exited container `%s`" % container.name)
-            container.remove()
-
-
-def start():
-    logging.info("Starting Orchest...")
-
-    # check if all images are present
-    if install_complete():
-        # start by cleaning up old containers lingering
-        clean_containers()
-
-        # start containers from CONTAINER_MAPPING that haven't started
-        running_containers = docker_client.containers.list()
-
-        # NOTE: is the repo tag always the first tag in the Docker Engine API?
-        running_container_images = [
-            running_container.image.tags[0]
-            for running_container in running_containers
-            if len(running_container.image.tags) > 0
-        ]
-
-        images_to_start = [
-            image_name
-            for image_name in CONTAINER_MAPPING.keys()
-            if image_name not in running_container_images
-        ]
-
-        for container_image in CONTAINER_MAPPING.keys():
-
-            if container_image in images_to_start:
-
-                container_spec = CONTAINER_MAPPING[container_image]
-
-                mounts = []
-                if 'mounts' in container_spec:
-                    mounts = [
-                        Mount(
-                            target=mount['target'],
-                            source=mount['source'],
-                            type='bind'
-                        )
-                        for mount in container_spec['mounts']
-                    ]
-
-                environment = container_spec.get('environment', {})
-                ports = container_spec.get('ports', {})
-                hostname = container_spec.get('hostname')
-                command = container_spec.get('command')
-
-                logging.info("Starting image %s" % container_image)
-
-                docker_client.containers.run(
-                    image=container_image,
-                    command=command,
-                    name=container_spec['name'],
-                    detach=True,
-                    mounts=mounts,
-                    network=DOCKER_NETWORK,
-                    environment=environment,
-                    ports=ports,
-                    hostname=hostname,
-                )
-
-        log_server_url()
-
-    else:
-        logging.info("Installation required. Starting installer.")
-        install_images()
-        install_network()
-        logging.info("Installation finished. Attempting to start...")
-        start()
-
-
-def install_network():
-    try:
-        docker_client.networks.get(DOCKER_NETWORK)
-    except docker.errors.NotFound as e:
-
-        logging.info("Orchest sends an anonymized ping to analytics.orchest.io. You can disable this by adding { \"TELEMETRY_DISABLED\": true } to config.json in %s" % HOST_CONFIG_DIR)
-
-        logging.info("Docker network %s doesn't exist: %s. Creating it." % (DOCKER_NETWORK, e))
-        # Create Docker network named "orchest" with a custom subnet such that
-        # containers can be spawned at custom static IP addresses.
-        ipam_pool = docker.types.IPAMPool(subnet='172.31.0.0/16')
-        ipam_config = docker.types.IPAMConfig(pool_configs=[ipam_pool])
-        docker_client.networks.create(
-            DOCKER_NETWORK,
-            driver="bridge",
-            ipam=ipam_config
-        )
-
-
-def install_complete():
-    missing_images = check_images()
-
-    if len(missing_images) > 0:
-        logging.warning("Missing images: %s" % missing_images)
-        return False
-
-    try:
-        docker_client.networks.get(DOCKER_NETWORK)
-    except docker.errors.NotFound as e:
-        logging.warning("Docker network (%s) not installed: %s" % (DOCKER_NETWORK, e))
-        return False
-
-    return True
-
-
-def check_images():
-    missing_images = []
-
-    for image in IMAGES:
-        try:
-            docker_client.images.get(image)
-            # logging.info("Image `%s` is installed." % image)
-        except docker.errors.ImageNotFound as e:
-            missing_images.append(image)
-        except docker.errors.APIError as e:
-            raise e
-
-    return missing_images
-
-
-def install_images():
-    for image in IMAGES:
-        try:
-            docker_client.images.get(image)
-        except docker.errors.ImageNotFound as e:
-            logging.info("Pulling image `%s` ..." % image)
-            docker_client.images.pull(image)
-            logging.info("Pulled image `%s`." % image)
-
-
-def get_application_url():
-    try:
-        # raise Exception if not found
-        docker_client.containers.get("orchest-webserver")
-    except Exception as e:
-        print(e)
-        return ""
-
-    port = CONTAINER_MAPPING["orchestsoftware/nginx-proxy:latest"]["ports"]["80/tcp"]
-    return "http://localhost:%i" % port
-
-
-def help_func():
-    for key in VALID_COMMANDS:
-        print("{0:20}\t {1}".format(key, VALID_COMMANDS[key]), flush=True)
-
-
-def stop():
-    # shut down containers
-    running_containers = docker_client.containers.list()
-
-    container_names = [CONTAINER_MAPPING[container_key]['name'] for container_key in CONTAINER_MAPPING]
-
-    for running_container in running_containers:
-        if len(running_container.image.tags) > 0 and running_container.image.tags[0] in IMAGES:
-
-            # don't kill orchest-ctl itself
-            if running_container.image.tags[0] == "orchestsoftware/orchest-ctl:latest":
-                continue
-
-            logging.info("Killing container %s" % running_container.name)
-            try:
-                running_container.kill()
-                running_container.remove()
-            except Exception as e:
-                print(e)
-        elif running_container.name in container_names:
-            logging.info("Killing container %s" % running_container.name)
-            try:
-                running_container.kill()
-                running_container.remove()
-            except Exception as e:
-                print(e)
-
-
-def log_server_url():
-    orchest_url = get_application_url()
-    if len(orchest_url) > 0:
-        logging.info("Orchest is running at: %s" % orchest_url)
-    else:
-        logging.warning("Orchest is not running.")
-
-
-def dev_mount_inject():
-    """Mounts to run Orchest in DEV mode.
-
-    The code is mounted at the correct locations inside the containers,
-    so that you changes you make to the code are directly reflected in
-    the application.
-
-    """
-    logging.info("Orchest starting in DEV mode. This mounts host directories "
-                 "to monitor for source code changes.")
-
-    # orchest-webserver
-    orchest_webserver_spec = CONTAINER_MAPPING["orchestsoftware/orchest-webserver:latest"]
-    orchest_webserver_spec['mounts'] += [
-        {
-            "source": os.path.join(
-                os.environ.get("HOST_PWD"),
-                "orchest",
-                "orchest-webserver",
-                "app"),
-            "target": "/orchest/orchest/orchest-webserver/app"
-        }
-    ]
-
-    orchest_webserver_spec['environment']["FLASK_APP"] = "main.py"
-    orchest_webserver_spec['environment']["FLASK_DEBUG"] = "1"
-    orchest_webserver_spec['command'] = [
-       "flask",
-       "run",
-       "--host=0.0.0.0",
-       "--port=80"
-    ]
-
-    # auth-server
-    orchest_auth_server_spec = CONTAINER_MAPPING["orchestsoftware/auth-server:latest"]
-    orchest_auth_server_spec['mounts'] += [
-        {
-            "source": os.path.join(
-                os.environ.get("HOST_PWD"),
-                "orchest",
-                "auth-server",
-                "app"),
-            "target": "/orchest/orchest/auth-server/app"
-        }
-    ]
-
-    orchest_auth_server_spec['environment']["FLASK_APP"] = "main.py"
-    orchest_auth_server_spec['environment']["FLASK_DEBUG"] = "1"
-    orchest_auth_server_spec['command'] = [
-       "flask",
-       "run",
-       "--host=0.0.0.0",
-       "--port=80"
-    ]
-
-    # orchest-api
-    orchest_api_spec = CONTAINER_MAPPING["orchestsoftware/orchest-api:latest"]
-    orchest_api_spec["mounts"] += [
-        {
-            "source": os.path.join(
-                os.environ.get("HOST_PWD"),
-                "orchest",
-                "orchest-api",
-                "app",
-                "app"),
-            "target": "/orchest/orchest/orchest-api/app/app"
-        }
-    ]
-    orchest_api_spec["ports"] = {
-        "80/tcp": 8080
-    }
-    orchest_api_spec["environment"]["FLASK_APP"] = "main.py"
-    orchest_api_spec["environment"]["FLASK_ENV"] = "development"
-    orchest_api_spec["command"] = [
-       "flask",
-       "run",
-       "--host=0.0.0.0",
-       "--port=80"
-    ]
-
-    # nginx-proxy
-    nginx_proxy_spec = CONTAINER_MAPPING["orchestsoftware/nginx-proxy:latest"]
-    nginx_proxy_spec['ports'] = {
-        "80/tcp": 80,
-        "443/tcp": 443
-    }
-
-
-def status():
-    running_containers = docker_client.containers.list()
-
-    orchest_container_names = [
-        CONTAINER_MAPPING[container_key]['name']
-        for container_key in CONTAINER_MAPPING
-    ]
-
-    running_prints = ['']
-    not_running_prints = ['']
-
-    for container in running_containers:
-        if container.name in orchest_container_names:
-            running_prints.append("Container %s running." % container.name)
-            orchest_container_names.remove(container.name)
-
-    for container_name in orchest_container_names:
-        not_running_prints.append("Container %s not running." % container_name)
-
-    if len(running_prints) > 1:
-        logging.info('\n'.join(running_prints))
-
-    if len(not_running_prints) > 1:
-        logging.info('\n'.join(not_running_prints))
-
-
-def init_logger():
-    logging.basicConfig(level=logging.INFO)
-
-    root = logging.getLogger()
-    if len(root.handlers) > 0:
-        h = root.handlers[0]
-        root.removeHandler(h)
-
-    formatter = logging.Formatter(logging.BASIC_FORMAT)
-    handler = logging.StreamHandler(sys.stdout)
-    handler.setFormatter(formatter)
-    root.addHandler(handler)
-
-
-def update():
-    logging.info("Updating Orchest...")
-
-    for image in IMAGES:
-        try:
-            logging.info("Pulling image `%s` ..." % image)
-            docker_client.images.pull(image)
-            logging.info("Pulled image `%s`." % image)
-        except Exception as e:
-            logging.error("Something went wrong while pulling image "
-                          "%s error: %s" % (image, e))
+import cmdline
+import config
+import utils
 
 
 def main():
-    init_logger()
+    utils.init_logger()
 
-    command_to_func = {
-        "start": start,
-        "help": help_func,
-        "stop": stop,
-        "status": status,
-        "update": update,
-    }
+    available_cmds = cmdline.get_available_cmds()
+    cmd_to_func = {cmd: getattr(cmdline, cmd) for cmd in available_cmds}
 
-    # default command
-    command = "help"
-
-    if len(sys.argv) > 1:
+    # Run the appropriate command. If an invalid command is given, then
+    # the `default_cmd` is run.
+    default_cmd = "help"
+    if len(sys.argv) <= 1:
+        command = default_cmd
+    else:
         command = sys.argv[1]
 
-    if len(sys.argv) > 2:
-        if sys.argv[2] == "dev":
-            dev_mount_inject()
-
-    if command not in VALID_COMMANDS.keys():
+    if command not in available_cmds:
         logging.error("Command `%s` is not supported." % command)
-        help_func()
-        return
+        command = default_cmd
+    else:
+        if len(sys.argv) > 2:
+            if sys.argv[2] == "dev":
+                config.RUN_MODE = "dev"
 
-    command_to_func[command]()
+    cmd_to_func[command]()
 
 
 if __name__ == '__main__':
